+++ src/main/java/cn/gudqs7/plugins/common/resolver/structure/StructureAndCommentResolver.java
@@ -415,8 +415,6 @@ public class StructureAndCommentResolver implements IStructureAndCommentResolver
                 }
             }
 
-        } else {
-            System.out.println(psiFieldType.getPresentableText() + " ==> not basic type, not ReferenceType");
         }
         return structureAndCommentInfo;
     }